+++ app/models/fetch_youtube_videos.rb
@@ -1,5 +1,4 @@
 require 'google/apis/youtube_v3'
-require 'google/apis/analytics_v3'
 
 class FetchYoutubeVideos
   # def initialize(session)
@@ -23,20 +22,4 @@ class FetchYoutubeVideos
        }
     end
   end
-end
-
-class FetchYoutubeAnalytics
-  
-  def self.call(session)
-    @auth_client ||= Signet::OAuth2::Client.new(session.credentials)
-    analytics ||= Google::Apis::AnalyticsV3::AnalyticsService.new   
-    analytics.get_realtime_data(
-      "ga:56789", # Table ID
-      "rt:activeUsers", # Metrics
-      options: { authorization: @auth_client}
-      )
-      # ERROR:
-      # Google::Apis::ClientError: insufficientPermissions: Request had insufficient authentication scopes.
-      # YA SE AMPLIARON LOS SCOPES Y ME PIDIERON LOS PERMISOS
-    end
 end
